+++ age/src/format.rs
@@ -1,5 +1,9 @@
 //! The age message format.
 
+use rand::{
+    distributions::{Distribution, Uniform},
+    thread_rng, RngCore,
+};
 use std::io::{self, Read, Write};
 
 use crate::primitives::HmacWriter;
@@ -55,6 +59,42 @@ impl From<plugin::RecipientLine> for RecipientLine {
     fn from(line: plugin::RecipientLine) -> Self {
         RecipientLine::Plugin(line)
     }
+}
+
+/// Creates a random recipient stanza that exercises the joint in the age v1 format.
+pub(crate) fn oil_the_joint() -> RecipientLine {
+    // Generate arbitrary strings between 1 and 9 characters long.
+    fn gen_arbitrary_string<R: RngCore>(rng: &mut R) -> String {
+        let length = Uniform::from(1..9).sample(rng);
+        Uniform::from(33..=126)
+            .sample_iter(rng)
+            .map(char::from)
+            .take(length)
+            .collect()
+    }
+
+    let mut rng = thread_rng();
+
+    // Add a prefix to the random tag so users know what is going on.
+    let tag = format!("joint-oil-{}", gen_arbitrary_string(&mut rng));
+
+    // Between this and the above generation bounds, the first line of the recipient
+    // stanza will be between five and 69 characters.
+    let args = (0..Uniform::from(0..5).sample(&mut rng))
+        .map(|_| gen_arbitrary_string(&mut rng))
+        .collect();
+
+    // A length between 0 and 100 bytes exercises the following stanza bodies:
+    // - Empty
+    // - Single short-line
+    // - Single full-line
+    // - Two lines, second short
+    // - Two lines, both full
+    // - Three lines, last short
+    let mut body = vec![0; Uniform::from(0..100).sample(&mut rng)];
+    rng.fill_bytes(&mut body);
+
+    plugin::RecipientLine { tag, args, body }.into()
 }
 
 pub struct HeaderV1 {